Pull Request for Release 11.0.0

+++ catalogue-service-micro/src/main/java/eu/nimble/service/catalogue/impl/CatalogueController.java
@@ -76,17 +76,18 @@ public class CatalogueController {
             @ApiResponse(code = 404, message = "No default catalogue for the party"),
             @ApiResponse(code = 500, message = "Failed to get CataloguePaginationResponse for the party")
     })
-    @RequestMapping(value = "/catalogue/{partyId}/pagination/{catalogueId}",
+    @RequestMapping(value = "/catalogue/{partyId}/pagination/{catalogueId:.+}",
             produces = {"application/json"},
             method = RequestMethod.GET)
     public ResponseEntity getDefaultCataloguePagination(@ApiParam(value = "Identifier of the party for which the catalogue to be retrieved", required = true) @PathVariable String partyId,
                                                         @ApiParam(value = "Identifier of the catalogue", required = true) @PathVariable String catalogueId,
-                                                        @ApiParam(value = "Number of catalogue lines to be included in CataloguePaginationResponse ") @RequestParam(value = "limit",required = true) Integer limit,
-                                                        @ApiParam(value = "Offset of the first catalogue line among all catalogue lines of the default catalogue for the party") @RequestParam(value = "offset",required = true) Integer offset,
+                                                        @ApiParam(value = "Number of catalogue lines to be included in CataloguePaginationResponse ",required = true) @RequestParam(value = "limit",required = true) Integer limit,
+                                                        @ApiParam(value = "Offset of the first catalogue line among all catalogue lines of the default catalogue for the party",required = true) @RequestParam(value = "offset",required = true) Integer offset,
                                                         @ApiParam(value = "Text to be used to filter the catalogue lines.Item name and description will be searched for the given text.") @RequestParam(value = "searchText",required = false) String searchText,
                                                         @ApiParam(value = "Identifier for the language of search text such as en and tr") @RequestParam(value = "languageId",required = false) String languageId,
                                                         @ApiParam(value = "Name of the category which is used to filter catalogue lines.Catalogue lines are added to the response if and only if they contain the given category.") @RequestParam(value = "categoryName",required = false) String categoryName,
                                                         @ApiParam(value = "Option used to sort catalogue lines") @RequestParam(value = "sortOption",required = false) CatalogueLineSortOptions sortOption,
+                                                        @ApiParam(value = "CatalogueUUID") @RequestParam(value = "catalogueUUId",required = false) String catalogueUUId,
                                                         @ApiParam(value = "The Bearer token provided by the identity service", required = true) @RequestHeader(value = "Authorization", required = true) String bearerToken) {
         log.info("Incoming request to get CataloguePaginationResponse for party: {}, catalogue id: {} with limit: {}, offset: {}", partyId, catalogueId, limit, offset);
         ResponseEntity tokenCheck = eu.nimble.service.catalogue.util.HttpResponseUtil.checkToken(bearerToken);
@@ -102,7 +103,7 @@ public class CatalogueController {
         CataloguePaginationResponse cataloguePaginationResponse;
 
         try {
-            cataloguePaginationResponse = service.getCataloguePaginationResponse(catalogueId, partyId,categoryName,searchText,languageId,sortOption,limit,offset);
+            cataloguePaginationResponse = service.getCataloguePaginationResponse(catalogueId, partyId,categoryName,searchText,languageId,sortOption,limit,offset,catalogueUUId);
         } catch (Exception e) {
             return createErrorResponseEntity(String.format("Failed to get CataloguePaginationResponse for party id: %s catalogue id: %s", partyId, catalogueId), HttpStatus.INTERNAL_SERVER_ERROR, e);
         }
@@ -437,13 +438,15 @@ public class CatalogueController {
                 ids = CataloguePersistenceUtil.getCatalogueIdListsForParty(partyId);
             }
 
-            for (String id : ids) {
-                service.deleteCatalogue(id, partyId);
-                Map<String,String> paramMap = new HashMap<String, String>();
-                paramMap.put("activity", CatalogueEvent.CATALOGUE_DELETE.getActivity());
-                paramMap.put("catalogueId", id);
-                paramMap.put("companyId", partyId);
-                LoggerUtils.logWithMDC(log, paramMap, LoggerUtils.LogLevel.INFO, "Successfully deleted catalogue, id: {} for company: {}", id, partyId);
+            if(ids != null){
+                for (String id : ids) {
+                    service.deleteCatalogue(id, partyId);
+                    Map<String,String> paramMap = new HashMap<String, String>();
+                    paramMap.put("activity", CatalogueEvent.CATALOGUE_DELETE.getActivity());
+                    paramMap.put("catalogueId", id);
+                    paramMap.put("companyId", partyId);
+                    LoggerUtils.logWithMDC(log, paramMap, LoggerUtils.LogLevel.INFO, "Successfully deleted catalogue, id: {} for company: {}", id, partyId);
+                }
             }
 
             log.info("Completed request to delete catalogues for party: {}, ids: {}, delete all: {}", partyId, idsLog, deleteAll);
@@ -804,49 +807,38 @@ public class CatalogueController {
     }
 
     @CrossOrigin(origins = {"*"})
-    @ApiOperation(value = "", notes = "Retrieves catalogue from id specific for a party.")
+    @ApiOperation(value = "", notes = "Retrieves the catalogue id's for a party.")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "Retrieved the catalogue successfully"),
+            @ApiResponse(code = 200, message = "Retrieved the catalogue uuid's successfully"),
             @ApiResponse(code = 401, message = "Invalid token. No user was found for the provided token"),
-            @ApiResponse(code = 404, message = "No catalogue found for the given id and party id"),
+            @ApiResponse(code = 404, message = "No catalogue uuids found for the given party"),
             @ApiResponse(code = 500, message = "Unexpected error while getting catalogue id's")
     })
-    @RequestMapping(value = "/catalogue/{partyId}/{id}/{standard}",
+    @RequestMapping(value = "/catalogue/idsuuids/{partyId}",
             produces = {"application/json"},
             method = RequestMethod.GET)
-    public ResponseEntity getCatalogueFromId(@ApiParam(value = "Identifier of the party for which the catalogue to be retrieved", required = true) @PathVariable String partyId,
-                                                        @ApiParam(value = "Identifier of the catalogue", required = true) @PathVariable String id,
-                                                        @ApiParam(value = "Data model standard that the provided catalogue is compatible with.", defaultValue = "ubl", required = true) @PathVariable String standard,
-                                                     @ApiParam(value = "The Bearer token provided by the identity service", required = true) @RequestHeader(value = "Authorization", required = true) String bearerToken) {
-
-        log.info("Incoming request to get catalogue for id: {} for party: {}",id,partyId);
+    public ResponseEntity getAllCatalogueIdsUUIDsForParty(@ApiParam(value = "Identifier of the party for which the catalogue to be retrieved", required = true) @PathVariable String partyId,
+            @ApiParam(value = "The Bearer token provided by the identity service", required = true) @RequestHeader(value = "Authorization", required = true) String bearerToken) {
+        log.info("Incoming request to get catalogue uuidid list for party: {}", partyId);
         ResponseEntity tokenCheck = eu.nimble.service.catalogue.util.HttpResponseUtil.checkToken(bearerToken);
-
         if (tokenCheck != null) {
             return tokenCheck;
         }
 
-        Configuration.Standard std;
+        List<Object[]> catalogueIds;
         try {
-            std = getStandardEnum(standard);
+            catalogueIds = service.getCatalogueIdAndNameForParty(partyId);
         } catch (Exception e) {
-            return createErrorResponseEntity("Invalid standard: " + standard, HttpStatus.BAD_REQUEST, e);
+            return createErrorResponseEntity("Failed to get catalogues for party id: " + partyId, HttpStatus.INTERNAL_SERVER_ERROR, e);
         }
 
-        Object catalogue;
-        try {
-            catalogue = service.getCatalogue(id,partyId,std);
-        } catch (Exception e) {
-            return createErrorResponseEntity("Failed to get catalogue for party id: " + partyId, HttpStatus.INTERNAL_SERVER_ERROR, e);
+        if (catalogueIds == null) {
+            log.info("No default catalogue for party: {}", partyId);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(String.format("No default catalogue for party: %s", partyId));
         }
 
-        if (catalogue == null) {
-            log.info("No catalogue for id: {}", id);
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(String.format("No catalogue for id: %s", id));
-        }
-
-        log.info("Completed request to get catalogue for party: {} and id {}", partyId,id);
-        return ResponseEntity.ok(catalogue);
+        log.info("Completed request to get catalogue uuid list for party: {}", partyId);
+        return ResponseEntity.ok(catalogueIds);
     }
 
     private ResponseEntity createErrorResponseEntity(String msg, HttpStatus status, Exception e) {
